Decouple regression tests from logging logic
Performance logging logic (log files, log directory creatation) is now
completely part of the performance logging configuration. More
specifically, the performance log directory and files are only relevant
to the `filelog` backend logging handler. For this reason, the logic of
creating the logging prefix is now removed from the runtime resources as
well.