Build systems infrastructure
The compilation phase is now separated from the programming environment.
It is part of the new concept of build systems. A build system is
responsible for generating the required commands for compiling a
code. The framework then uses these commands to generate a build job
script which is then submitted. Currently, only local compilation is
supported.

The current behavior of the `sourcepath` and `sourcesdir` attributes is
maintained for convenience. Internally, they translate to concrete build
systems that are set up accordingly.

All build systems share some basic attributes and behavior:

- The compilers and compilation flags. By default, if not specified the
  corresponding values from the current programming environment will be
  used.
- The ability to ignore completely the current programming
  environment. A build system may be configured independently of the
  current programming environment by explicitly setting the compilers
  and compilation flags.

The above design allows the programming environment to become immutable
holding the global default values for each system.
Currently, for backward compatibility, it is not yet immutable, but
setting its attributes is now deprecated.

Two build systems are provided by this commit:

1. `SingleSource`: This build system is responsible for compiling a
    single source file in any of the recognized programming languages,
    i.e., C, C++, Fortran and CUDA.

2. `Make`: This build system is responsible for compiling a project
   using the `make` command.