Add option to increase output verbosity level
This PR adds the option to increase the verbosity level of the
PrettyPrinter. However, there is only one call to printer.debug and
there is on call to printer.verbose in the entire code. The former is
related to a module not being properly loaded, which is difficult to
reproduce, since we do have correct checks on ReFrame.

I am not sure if it is part of the issue to also add the printer.debug
and printer.verbose calls into the code.