Move checking of the perf var extracted variable inside the pipeline
The reason for that is that we don't do any argument type checking in any other
sanity function, so it would be strange to do it in either `assert_reference()`
or `assert_bounded()` only. The reason for that is that type and value errors
are handled specially if they are raised from user test context. In this
particular case, though, the `assert_reference()` is called from the framework
context as a result of user input. So it makes more sense then to do the type
checking inside `check_performance()`, where we call it.

We also don't type check the reference and lower/upper thresholds, because this
should be addressed correctly by the updated declaration of the `reference`
field as soon as https://github.com/eth-cscs/reframe/issues/1147 is fixed (see
also related FIXME note for the `reference` field).

Finally, this commit adds also a pipeline unit test to check for this condition.