Add a powerful repr() function for debugging
This function replaces the older `debug.repr()` and uses `util.ppretty()` for
nicely printing the object attributes. `util.ppretty()` is also extended in
order to recursively call the new `repr()` function. Objects are printed only
once by this new `repr()` avoiding infinite recursions.

Finally, we use that function to print the whole ReFrame runtime state in case
issue #1369 appears.

This function does not take into account the `__slots__` for the moment.
In cases, where an object does not have a `__dict__` attribute, the builtin
`repr()` will be used.