Revisit Graylog log handler as well as the dynamic field formatting
The log record extras are not formatted at all when they are defined in the
`LoggerAdapter`'s extras. They can be raw objects. Our custom formatter will
format them, but it will NOT overwrite the log record attributes with their
formatted values, so that the record contains all the information. The builtin
formatters can still work with that, except that they will use the `str()`
method to format objects.

The Graylog handler does not use a formatter by design. Instead, it encodes the
whole record object into JSON and sends it over to the Graylog server. All we do
in this case is to use our own JSON encoder, which can recognize the ReFrame
objects.