Add nomod to example configuration file
A laptop would generally not have modules, so saying that explicitly
will help highlight to new users that a module setting can be
useful. There is a modules setting below for daint, but at least one
user skimmed over the more complex case.

The line numbers for the literal includes need to be bumped if they
are greater than 15, and a range including 15 is shown, then the
highlighting can need a bump. Note that the latter index within the
range shown, so don't always need a bump.

Fixes #1887