Reimplement pipeline hooks
Pipeline hooks are now implemented in the metaclass.
The logic is very similar to that of the variables and parameters.
As the class hierarchy is built, a `HookRegistry` is maintained containing all
the hooks. When the test object is created for the first time, all of its
pipeline methods are wrapped with the corresponding pipeline hooks.

Moving the implementation to the metaclasses brings a set of advantages:

- Inheritance is handled by Python and there is no need to manually ascend the
  class hierarchy when the hook is applied.
- For storing the hooks of each phase, we use an ordered set and we make sure
  that firstly the hooks of the class that is being currently created are
  inserted and then those of its parents. This way we ensure that hooks are
  properly overriden.
- By using metaclasses, hooks are more consistent and are automatically applied
  correctly to any method overrides.