Reimplement the pipeline hook mechanism
Pipeline hooks are no more applied to the pipeline functions with the decorator.
This has the following advantages:

a. Hooks will be automatically applied to any pipeline function that is
   overriden by a subclass.
b. The hooks will wrap the leaf function and not the call to
  `super().pipeline_fn()`
c. The hooks will be applied only once in all cases even if a test and its child
   classes are instantiated multiple times.