Improve implementation of `Bool` type.
`Bool` type no more holds a value. This is not relevant, because it is supposed
to encode a type. It's only responsibility is to take care of the special
conversion from string to booleans. There is no need also to support conversion
from any other type. If the type is used in a `TypedField`, assignments from
booleans would work, since `bool` is registered as a subclass of `Bool`, and any
other non-string assignment would fail due to incompatible type.