Introduce variable aliases
This practically extends on the field aliasing that we used to have and brings
the same concept to variables. Users can now define a variable as an alias of
another one. This is quite useful for transitional periods and deprecations when
a variable is renamed.

This commit also introduces implementation improvements on how variables are
inherited.