Add ssh_audit test to hpctestlist
The test downloads the ssh-audit tool from GitHub and checks if there is
any configuration that is considered as a failure.

The test is compatible with all the job schedulers. Because it only
checks the localhost, where the test script runs. So, it becomes the
responsibility of ReFrame to spawn the test script in the desired
systems.