display python imports in first example in tutorial
The tutorial was skipping the two imports

    import reframe as rfm
    import reframe.utility.sanity as sn

by using the :pyobject: option to only highlight stream_test. This is
the first example shown in the tutorial and it is important that the
entire file is displayed.

Presumably this technique is used to avoid the license notice; we can
use :lines: instead.